% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tables.R
\name{clean_tables}
\alias{clean_tables}
\title{Clean tables}
\usage{
clean_tables(df, section, subsection, geography)
}
\arguments{
\item{df}{Data fram}

\item{section}{Section}

\item{subsection}{Subsection}

\item{geography}{Geography}
}
\description{
Clean ECLDS tables
}
