% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{List of counties in Minnesota}
\format{
A data frame
}
\usage{
counties
}
\description{
A dataset containing the list of counties (names and id's).
}
\keyword{datasets}
