% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_remotedriver.R
\name{create_remotedriver}
\alias{create_remotedriver}
\title{Create Remote Driver}
\usage{
create_remotedriver(browser = "firefox", port = 4445)
}
\arguments{
\item{browser}{Browser (default is set to 'firefox')}

\item{port}{Port (default is set to 4445)}
}
\description{
Create remote driver for navigation
}
