% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disabilities}
\alias{disabilities}
\title{List of disabilities available in the website}
\format{
A data frame
}
\usage{
disabilities
}
\description{
A dataset containing the list of disabilities available in the website
}
\keyword{datasets}
