% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables.R
\name{export_eclds_all}
\alias{export_eclds_all}
\title{Export ECLDS data}
\usage{
export_eclds_all(data, path)
}
\arguments{
\item{data}{ECLDS data acquired using \code{scrape_eclds_all}.}

\item{path}{Path of the folder to write the files}
}
\description{
Export ECLDS data coming from scraping all (main) tables to data files
}
