% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{languages}
\alias{languages}
\title{List of languages available in the website}
\format{
A data frame
}
\usage{
languages
}
\description{
A dataset containing the list of languages available in the website
}
\keyword{datasets}
