% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_districts}
\alias{school_districts}
\title{List of school districts in Minnesota}
\format{
A data frame
}
\usage{
school_districts
}
\description{
A dataset containing the list of school districts (names and id's).
}
\keyword{datasets}
