% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_eclds.R
\name{scrape_eclds}
\alias{scrape_eclds}
\title{Scrape ECLDS pages}
\usage{
scrape_eclds(
  section,
  subsection,
  geography,
  year,
  browser = "firefox",
  port = 4445
)
}
\arguments{
\item{section}{"kindergarten" or "birth to prek"}

\item{subsection}{"early care and education", "child demographics", "family demographics", "economic and food assistance", "kindergarten attendance", "kindergarten summary" for kindergarten; "early childhood screening", "scholarships", "parent aware", "community assessment" for "birth to prek"}

\item{geography}{"county" or "school district"}

\item{year}{Vector of years. 2014 to 2020 as of September 2021.}

\item{browser}{Browser in the remote driver}

\item{port}{Port for the remote driver}
}
\description{
Scrape specific sections of the ECLDS.
}
